+++ src/observable/connectable.jl
@@ -1,6 +1,5 @@
-export ConnectableObservable, connectable, connect
+export connectable, connect
 
-import Base: ==
 import Base: show
 
 """
@@ -11,10 +10,12 @@ It subscribes to the source when its `connect` method is called.
 
 See also: [`connect`](@ref), [`Subscribable`](@ref)
 """
-@subscribable struct ConnectableObservable{D, J, S} <: Subscribable{D}
+struct ConnectableObservable{D, J, S} <: Subscribable{D}
     subject :: J
     source  :: S
 end
+
+Base.show(io::IO, ::ConnectableObservable{D}) where D = print(io, "ConnectableObservable($D)")
 
 function on_subscribe!(observable::ConnectableObservable, actor)
     return subscribe!(observable.subject, actor)
@@ -46,31 +47,19 @@ connect(c);
 
 ```
 
-See also: [`ConnectableObservable`](@ref), [`connect`](@ref), [`subscribe!`](@ref)
+See also: [`connect`](@ref), [`subscribe!`](@ref)
 """
-connectable(subject::J, source::S) where { J, S } = as_connectable(as_subject(J), as_subscribable(S), subject, source)
+connectable(subject, source) = as_connectable(eltype(subject), eltype(source), subject, source)
 
-as_connectable(::InvalidSubjectTrait, ::InvalidSubscribableTrait, subject, source) = throw(InvalidSubjectTraitUsageError(subject))
-as_connectable(::InvalidSubjectTrait, _,                          subject, source) = throw(InvalidSubjectTraitUsageError(subject))
-as_connectable(_,                     ::InvalidSubscribableTrait, subject, source) = throw(InvalidSubscribableTraitUsageError(source))
-
-as_connectable(::ValidSubjectTrait{D1}, ::SimpleSubscribableTrait{D2},    subject, source) where { D1, D2  } = throw(InconsistentActorWithSubscribableDataTypesError{D2, D1}(source, subject))
-as_connectable(::ValidSubjectTrait{D1}, ::ScheduledSubscribableTrait{D2}, subject, source) where { D1, D2  } = throw(InconsistentActorWithSubscribableDataTypesError{D2, D1}(source, subject))
-
-as_connectable(::ValidSubjectTrait{D1},  ::SimpleSubscribableTrait{D2},    subject::J, source::S) where { D1, D2 <: D1, J, S } = ConnectableObservable{D2, J, S}(subject, source)
-as_connectable(::ValidSubjectTrait{D1},  ::ScheduledSubscribableTrait{D2}, subject::J, source::S) where { D1, D2 <: D1, J, S } = ConnectableObservable{D2, J, S}(subject, source)
+as_connectable(::Type{D1}, ::Type{D2}, subject, source)       where { D1, D2 }             = error("Element types $(D!) and $(D2) of subject and source observables should match.")
+as_connectable(::Type{D1}, ::Type{D2}, subject::J, source::S) where { D1, D2 <: D1, J, S } = ConnectableObservable{D2, J, S}(subject, source)
 
 """
     connect(connectable::ConnectableObservable)
 
-When `connect` is called, the subject passed to the multicast operator is subscribed to the source and the subject’s observers receive the multicast notifications,
-which fits our basic mental model of stream multicasting. Returns a subscription.
+When `connect` is called, the subject passed to the multicast operator is subscribed to the source and the subject’s observers receive the multicast notifications.
+Returns a subscription.
 
 See also: [`connectable`](@ref), [`ConnectableObservable`](@ref)
 """
 connect(connectable::ConnectableObservable) = subscribe!(connectable.source, connectable.subject)
-
-Base.:(==)(c1::ConnectableObservable{D},  c2::ConnectableObservable{D})  where { D      } = c1.subject == c2.subject && c1.source == c2.source
-Base.:(==)(c1::ConnectableObservable{D1}, c2::ConnectableObservable{D2}) where { D1, D2 } = false
-
-Base.show(io::IO, ::ConnectableObservable{D}) where D = print(io, "ConnectableObservable($D)")