+++ src/operators/publish.jl
@@ -1,40 +1,37 @@
-export publish
-export publish_behavior
-export publish_replay
-export publish_recent
+export publish, publish_behavior, publish_replay, publish_recent
 
 """
-    publish(; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
+    publish(; scheduler = AsapScheduler())
 
 `publish()` is a shortcut for `multicast(SubjectFactory())`
 
-See also: [`AbstractOperator`](@ref), [`multicast`](@ref), [`SubjectFactory`](@ref)
+See also: [`Operator`](@ref), [`multicast`](@ref), [`SubjectFactory`](@ref)
 """
-publish(; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler } = multicast(SubjectFactory(scheduler))
+publish(; scheduler = AsapScheduler()) = multicast(SubjectFactory(scheduler))
 
 """
-    publish_behavior(default; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
+    publish_behavior(default; scheduler = AsapScheduler())
 
 `publish_behavior(default)` is a shortcut for `multicast(BehaviorSubjectFactory(default))`
 
-See also: [`AbstractOperator`](@ref), [`multicast`](@ref), [`BehaviorSubjectFactory`](@ref)
+See also: [`Operator`](@ref), [`multicast`](@ref), [`BehaviorSubjectFactory`](@ref)
 """
-publish_behavior(default; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler } = multicast(BehaviorSubjectFactory(default, scheduler = scheduler))
+publish_behavior(default; scheduler = AsapScheduler()) = multicast(BehaviorSubjectFactory(default, scheduler = scheduler))
 
 """
-    publish_replay(size::Int; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
+    publish_replay(size::Int; scheduler = AsapScheduler())
 
 `publish_replay(size)` is a shortcut for `multicast(ReplaySubjectFactory(size))`
 
-See also: [`AbstractOperator`](@ref), [`multicast`](@ref), [`ReplaySubjectFactory`](@ref)
+See also: [`Operator`](@ref), [`multicast`](@ref), [`ReplaySubjectFactory`](@ref)
 """
-publish_replay(size::Int; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler } = multicast(ReplaySubjectFactory(size, scheduler = scheduler))
+publish_replay(size::Int; scheduler = AsapScheduler()) = multicast(ReplaySubjectFactory(size, scheduler = scheduler))
 
 """
-    publish_recent(; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
+    publish_recent(; scheduler = AsapScheduler())
 
 `publish_recent(size)` is a shortcut for `multicast(RecentSubjectFactory())`
 
-See also: [`AbstractOperator`](@ref), [`multicast`](@ref), [`RecentSubjectFactory`](@ref)
+See also: [`Operator`](@ref), [`multicast`](@ref), [`RecentSubjectFactory`](@ref)
 """
-publish_recent(; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler } = multicast(RecentSubjectFactory(scheduler = scheduler))
+publish_recent(; scheduler = AsapScheduler()) = multicast(RecentSubjectFactory(scheduler = scheduler))