+++ src/operators/skip_next.jl
@@ -16,7 +16,7 @@ Stream of type `<: Subscribable{L}` where `L` refers to type of source stream
 ```jldoctest
 using Rocket
 
-source = from([ 1, 2, 3 ])
+source = from_iterable([ 1, 2, 3 ])
 subscribe!(source |> skip_next(), logger())
 ;
 
@@ -25,30 +25,34 @@ subscribe!(source |> skip_next(), logger())
 
 ```
 
-See also: [`AbstractOperator`](@ref), [`InferableOperator`](@ref), [`ProxyObservable`](@ref), [`skip_error`](@ref), [`skip_complete`](@ref), [`logger`](@ref)
+See also: [`Operator`](@ref), [`skip_error`](@ref), [`skip_complete`](@ref), [`logger`](@ref)
 """
 skip_next() = SkipNextOperator()
 
-struct SkipNextOperator <: InferableOperator end
+struct SkipNextOperator <: Operator end
 
-function on_call!(::Type{L}, ::Type{L}, operator::SkipNextOperator, source) where L
-    return proxy(L, source, SkipNextProxy())
+operator_eltype(::SkipNextOperator, ::Type{L}) where L = L
+
+struct SkipNextSubscribable{L, S} <: Subscribable{L}
+    source :: S
 end
 
-operator_right(operator::SkipNextOperator, ::Type{L}) where L = L
-
-struct SkipNextProxy <: ActorProxy end
-
-actor_proxy!(::Type{L}, proxy::SkipNextProxy, actor::A) where { L, A } = SkipNextActor{L, A}(actor)
-
-struct SkipNextActor{L, A} <: Actor{L}
+struct SkipNextActor{A}
     actor :: A
 end
 
-on_next!(actor::SkipNextActor{L}, data::L) where L = begin end
-on_error!(actor::SkipNextActor, err)               = error!(actor.actor, err)
-on_complete!(actor::SkipNextActor)                 = complete!(actor.actor)
+function on_call!(::Type{L}, ::Type{L}, operator::SkipNextOperator, source::S) where { L, S }
+    return SkipNextSubscribable{L, S}(source)
+end
 
-Base.show(io::IO, ::SkipNextOperator)         = print(io, "SkipNextOperator()")
-Base.show(io::IO, ::SkipNextProxy)            = print(io, "SkipNextProxy($L)")
-Base.show(io::IO, ::SkipNextActor{L}) where L = print(io, "SkipNextActor($L)")
+function on_subscribe!(source::SkipNextSubscribable, actor::A) where A
+    return subscribe!(source.source, SkipNextActor{A}(actor))
+end
+
+on_next!(actor::SkipNextActor, data) = begin end
+on_error!(actor::SkipNextActor, err) = error!(actor.actor, err)
+on_complete!(actor::SkipNextActor)   = complete!(actor.actor)
+
+Base.show(io::IO, ::SkipNextOperator)                = print(io, "SkipNextOperator()")
+Base.show(io::IO, ::SkipNextSubscribable{L}) where L = print(io, "SkipNextProxy($L)")
+Base.show(io::IO, ::SkipNextActor)                   = print(io, "SkipNextActor()")