+++ test/operators/test_operator_tap.jl
@@ -9,8 +9,6 @@ include("../test_helpers.jl")
 
     println("Testing: operator tap()")
 
-    run_proxyshowcheck("Tap", tap((_) -> begin end))
-
     sideeffects1 = []
     sideeffects2 = []
     sideeffects3 = []
@@ -18,11 +16,11 @@ include("../test_helpers.jl")
 
     run_testset([
         (
-            source = from(1:5) |> tap_on_subscribe(() -> sideeffects1 = []) |> tap((d) -> push!(sideeffects1, d)),
+            source = from_iterable(1:5) |> tap_on_subscribe(() -> sideeffects1 = []) |> tap((d) -> push!(sideeffects1, d)),
             values = @ts([ 1:5, c ])
         ),
         (
-            source = from(1:5) |> tap_on_subscribe(() -> sideeffects2 = []) |> tap((d) -> push!(sideeffects2, d)) |> skip_next(),
+            source = from_iterable(1:5) |> tap_on_subscribe(() -> sideeffects2 = []) |> tap((d) -> push!(sideeffects2, d)) |> skip_next(),
             values = @ts(c)
         ),
         (